requirements.txt
amqp==5.2.0
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
kombu==5.3.4
Markdown==3.4.4
prompt-toolkit==3.0.43
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
django-cors-headers==4.2.0
psycopg==3.1.9
channels==4.0.0
channels-redis==4.1.0
daphne
drf-spectacular
pre-commit
black